review: cleanup+suggestions

+++ src/diffusers/pipelines/flux2/pipeline_flux2_klein_inpaint.py
@@ -326,7 +326,9 @@ class Flux2KleinInpaintPipeline(DiffusionPipeline, Flux2LoraLoaderMixin):
             coords = torch.cartesian_prod(t, h, w, l)
             out_ids.append(coords)
 
-        return torch.stack(out_ids)
+        # Cast position ids to float32: these are RoPE coordinate indices and the Neuron
+        # compiler does not support int64 tensors. float32 is exact for this index range.
+        return torch.stack(out_ids).float()
 
     @staticmethod
     # Copied from diffusers.pipelines.flux2.pipeline_flux2.Flux2Pipeline._prepare_latent_ids
@@ -359,7 +361,9 @@ class Flux2KleinInpaintPipeline(DiffusionPipeline, Flux2LoraLoaderMixin):
         # Expand to batch: (B, H*W, 4)
         latent_ids = latent_ids.unsqueeze(0).expand(batch_size, -1, -1)
 
-        return latent_ids
+        # Cast position ids to float32: these are RoPE coordinate indices and the Neuron
+        # compiler does not support int64 tensors. float32 is exact for this index range.
+        return latent_ids.float()
 
     @staticmethod
     def _prepare_image_ids(
